requirements.txt
torch
numpy
Pillow
tqdm
realesrgan
basicsr
opencv-python
